+++ common/src/main/java/com/gregtechceu/gtceu/core/mixins/GregTechMixinPlugin.java
@@ -25,12 +25,18 @@ public class GregTechMixinPlugin implements IMixinConfigPlugin {
             return MixinPluginShared.isClassFound("dev.latvian.mods.kubejs.KubeJSPlugin");
         } else if (mixinClassName.contains("com.gregtechceu.gtceu.core.mixins.create")) {
             return MixinPluginShared.isClassFound("com.simibubi.create.compat.Mods");
+        } else if (mixinClassName.contains("com.gregtechceu.gtceu.core.mixins.rei")) {
+            return MixinPluginShared.isClassFound("me.shedaniel.rei.api.common.plugins.REIPlugin");
         } else if (mixinClassName.contains("com.gregtechceu.gtceu.fabric.core.mixins.kjs")) {
             return MixinPluginShared.isClassFound("dev.latvian.mods.kubejs.fabric.KubeJSFabric");
         } else if (mixinClassName.contains("com.gregtechceu.gtceu.forge.core.mixins.kjs")) {
             return MixinPluginShared.isClassFound("dev.latvian.mods.kubejs.forge.KubeJSForge");
         } else if (mixinClassName.contains("com.gregtechceu.gtceu.core.mixins.top")) {
             return MixinPluginShared.isClassFound("mcjty.theoneprobe.api.ITheOneProbe");
+        } else if (mixinClassName.contains("com.gregtechceu.gtceu.forge.core.mixins.jei")) {
+            return MixinPluginShared.isClassFound("mezz.jei.api.IModPlugin");
+        }else if (mixinClassName.contains("com.gregtechceu.gtceu.forge.core.mixins.emi")) {
+            return MixinPluginShared.isClassFound("dev.emi.emi.api.EmiPlugin");
         }
         return true;
     }